requirements.txt
Django==1.11.1
django_redis==4.10.0
redis==3.1.0
djangorestframework==3.9.1
